requirements.txt
asgiref==3.5.2
Django==4.1.3
django-tempus-dominus==5.1.2.17
sqlparse==0.4.3
